pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "acortadorenlaces"
version = "0.1.0"
description = "Aplicación web para acortar URLs con soporte para registro de usuarios, autenticación y gestión de enlaces personalizados"
authors = [
    {name = "Juande Molina", email = "juande.molina@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "Flask",
    "Flask-SQLAlchemy",
    "Flask-Migrate",
    "Flask-Login",
    "python-dotenv",
    "Werkzeug",
    "psycopg2-binary",
    "pytest",
]

[project.urls]
Homepage = "https://github.com/juandemolina/acortadorenlaces"
Repository = "https://github.com/juandemolina/acortadorenlaces.git"
Issues = "https://github.com/juandemolina/acortadorenlaces/issues"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"